pyproject.toml
[tool.poetry]
name = "django_view_composer"
version = "0.1.0"
description = ""
authors = ["Josh Higgins <5124298+joshiggins@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
